# dump/golang/grpc/api/openapi/broker.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "broker.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Broker"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/mq.Broker/Publish": {
      "post": {
        "operationId": "Broker_Publish",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mqPublishResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mqPublishRequest"
            }
          }
        ],
        "tags": [
          "Broker"
        ]
      }
    },
    "/mq.Broker/Subscribe": {
      "post": {
        "operationId": "Broker_Subscribe",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/mqMessageResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of mqMessageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mqSubscribeRequest"
            }
          }
        ],
        "tags": [
          "Broker"
        ]
      }
    }
  },
  "definitions": {
    "mqMessageResponse": {
      "type": "object",
      "properties": {
        "body": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "mqPublishRequest": {
      "type": "object",
      "properties": {
        "topic": {
          "type": "string"
        },
        "body": {
          "type": "string",
          "format": "byte"
        },
        "ttl": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "mqPublishResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "mqSubscribeRequest": {
      "type": "object",
      "properties": {
        "topic": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}